Move files around and delete unused code in esm-patient-medications
