+++ packages/esm-patient-medications-app/src/active-medications/active-medications.scss
